+++ CMakeLists.txt
@@ -11,8 +11,12 @@ option(TEST "generate test targets" FALSE)
 add_library(tool-libs INTERFACE)
 target_include_directories(tool-libs INTERFACE .)
 
-add_subdirectory(stm)
-add_subdirectory(linux)
+if(STM32_TOOLCHAIN_PATH)
+    add_subdirectory(stm)
+else()
+    add_subdirectory(linux)
+endif()
+
 if (PROJECT_IS_TOP_LEVEL OR TEST)
     add_subdirectory(tests)
 endif()
