+++ linux/sys/host.h
