+++ stm/sys/hal.h
