+++ stm/sys/registry.h
