+++ stm/sys/spi.h
